+++ src/pages/ProfilePage/ProfileMainPage/styles.js
@@ -1,0 +1,12 @@
+import { makeStyles } from '@material-ui/core/styles'
+
+export const useProfilePageStyles = makeStyles(() => ({
+  root: {
+    flexGrow: 1,
+    marginTop: '5%',
+    marginBottom: '2%'
+  },
+  item: {
+    marginRight: '10%',
+  },
+}))
