+++ src/pages/ProfilePage/UpdateDataPage/index.js
@@ -1,0 +1,10 @@
+import React from 'react'
+
+const UpdateDataPage=()=>{
+
+  return (
+    <h2>Hola mundo</h2>
+  )
+}
+
+export default UpdateDataPage
